==> scripts/M1_Hydro_pop_patterns/3c_Grp_means_tests_ECOREG_LO_NLA12.R <==
###############
## Group means comparison tests
## For NLA 2012 continuous sample population values
##  SIZE-ADJUSTED (>=4ha lakes)
##
## Spoke with Tom Kincaid 9/21/17
## The NLA is not a equally probability dataset - so many standard stats tests are not appropriate
## Rather they hand calculate z-scores 
##  using the population mean estimates and standard error values from spsurvey

## z = (|mean1 - mean2|)/[sqrt{(stderror1 ^2)+ stderror2^2)}]
##  Look up p-value using pnorm function
##  Multiply by 2 to get two tail p-value 
## NOTE - may be doing multiple group comparisons - use a lower p significance value to be more conservative (e.g., 0.01)

# 2/2/18

# 3/14/18 - added test of just lake origin

# 4/12/18 - SCALED DRAWDOWN variables - use log transformed variables to do group means tests

# 5/14/18 - Using more complete dataset with NAs - increased number of lakes with E:I

# 5/15/18 - Using more complete dataset

# 8/6/18 - UPDATED - more E:I and RT estimates

# 7/9/19 - UPDATED DATASETS WITH NEW POPULATION WEIGHTS

# SUBPOPULTION = ECOREGION + LAKE ORIGIN
################

#################

rm(list=ls())

#############
## LOAD DATA - Percentile spsurvey output (from Continuous_variable_Estimates_EF_NLA07_07SEPT17.R)
#############
####################
## Create datasets from ouptput to make boxplots using percentile values
####################
library(reshape2)
library(dplyr)

# FULL 2012 dataset n=1038 sampled lakes
# 7/10/18
# 8/6/18


######################
## LOAD NLA 2012 SIZE ADJUSTED (n = 951)
#p<-read.csv("M:/Net MyDocuments/a_Water_Level/Analysis/NLA_weighted_calculations/R_output/NLA12/NLA12_CONTINUOUS_SIZEADJ_percentile_06AUG18.csv") #OLD p<-read.csv("M:/Net MyDocuments/a_Water_Level/Analysis/NLA_weighted_calculations/R_output/NLA12/NLA12_CONTINUOUS_percentile_10OCT17.csv")
p<-read_csv("Routput/pop_calculations/NLA12_CONTINUOUS_SIZEADJ_percentile_27JUN19.csv")
names(p)
table(p$Subpopulation)
table(p$Type)

############
## Subset data "Type" = "WSA9_by_Lake_Origin"
# 1/18/18
# 7/10/18 - includes scaled vertical (modified depth)
# 8/6/18 updated E:I
############
p1<-subset(p,Type=="WSA9_by_Lake_Origin")
table(p1$Type)
p1<-droplevels(p1)

# Need to modify Subpopulation to pull out grouping factors
p1$ECOREG_use<-substr(p1$Subpopulation,1,3)

p1$Lake_Origin_use<-substr(p1$Subpopulation,5,13)
table(p1$Lake_Origin_use)

#
p1$ECOREG_LO<-p1$Subpopulation
#Order Ecoregions west to east
p1$ECOREG_use <- ordered(p1$ECOREG_use, levels=c("WMT","XER","NPL","SPL","TPL","UMW","SAP","NAP","CPL"))

# Subset data by "Mean" Statistic 
# Pull out Mean statistic
p2 <-subset(p1,Statistic=="Mean")

#################
## Mean test function - from Tom 9/22/17, modified by Emi
#################

p2$Indicator <- as.character(p2$Indicator)
p2$ECOREG_use <-as.character(p2$ECOREG_use)
p2$Lake_Origin_use <- as.character(p2$Lake_Origin_use)
p2$ECOREG_LO <-as.character(p2$ECOREG_LO)

# Create a data frame for test results
Results <- data.frame(Indicator=NA,GROUP_1=NA,GROUP_2=NA,  # TWo ecoregion classes
                      NResp_1=NA, NResp_2=NA, Estimate_1=NA,
                      Estimate_2=NA, Difference=NA, StdError=NA, z_Score=NA, 
                      p_Value=NA)

# Calculate confidence bounds, zscores, and pvalues by looping through
# Indicators, Lake Origin, and Ecoregions

r<- 0
for(ind in unique(p2$Indicator)){
  #for(lk in unique(p2$Lake_Origin_use)){
  tempdf<-droplevels(subset(p2, Indicator==ind)) # & Lake_Origin_use ==lk
  if(nrow(tempdf)>1){
    nlev<- length(unique(tempdf$ECOREG_LO))
    for(i in 1:(nlev-1)){
      for(j in 2:nlev){
        r<-r+1
        diff<- tempdf$Estimate[i] - tempdf$Estimate[j]
        stderr<-sqrt(tempdf$StdError[i]^2 + tempdf$StdError[j]^2)
        tst<-abs(diff)/stderr
        pval<-2*(1-pnorm(tst))
        Results[r, 1:11] <- c(ind,tempdf$ECOREG_LO[i],tempdf$ECOREG_LO[j], #lk,
                              tempdf$NResp[i],
                              tempdf$NResp[j],tempdf$Estimate[i],
                              tempdf$Estimate[j], diff, stderr,tst,pval)
      }
    }
  }
}

# 7/9/19 - SIZE ADJUSTED NLA 2012 dataset
write.csv(Results,"C:/Users/EFergus/OneDrive - Environmental Protection Agency (EPA)/a_Water_Level/Analysis/NLA_weighted_calculations/R_output/NLA12/Group_means_tests_NLA12_SIZEADJ_LO_ECOREG_SCALED_09JUL19.csv" )


#########
## SUBSET DATA BY ECOREGION ONLY
#########
#############
# Subset by lake type and ecoregion
p1<-subset(p,Type=="WSA9_Ecoregions")
table(p1$Type)
p1<-droplevels(p1)

#Order Ecoregions west to east
p1$ECOREG_use <- ordered(p1$Subpopulation, levels=c("WMT","XER","NPL","SPL","TPL","UMW","SAP","NAP","CPL"))

table(p1$ECOREG_use)
table(p1$Indicator)

# subset data by central tendancy statistic
p2 <-subset(p1,Statistic=="Mean")

#################
## Mean test function - from Tom 9/22/17, modified by Emi
#################

p2$Indicator <- as.character(p2$Indicator)
p2$ECOREG_use <-as.character(p2$ECOREG_use)
#p2$Lake_Origin_use <- as.character(p2$Lake_Origin_use)
#p2$ECOREG_LO <-as.character(p2$ECOREG_LO)

# Create a data frame for test results
Results <- data.frame(Indicator=NA,GROUP_1=NA,GROUP_2=NA,  # TWo ecoregion classes
                      NResp_1=NA, NResp_2=NA, Estimate_1=NA,
                      Estimate_2=NA, Difference=NA, StdError=NA, z_Score=NA, 
                      p_Value=NA)

# Calculate confidence bounds, zscores, and pvalues by looping through
# Indicators, Lake Origin, and Ecoregions

r<- 0
for(ind in unique(p2$Indicator)){
  #for(lk in unique(p2$Lake_Origin_use)){
  tempdf<-droplevels(subset(p2, Indicator==ind)) # & Lake_Origin_use ==lk
  if(nrow(tempdf)>1){
    nlev<- length(unique(tempdf$ECOREG_use))
    for(i in 1:(nlev-1)){
      for(j in 2:nlev){
        r<-r+1
        diff<- tempdf$Estimate[i] - tempdf$Estimate[j]
        stderr<-sqrt(tempdf$StdError[i]^2 + tempdf$StdError[j]^2)
        tst<-abs(diff)/stderr
        pval<-2*(1-pnorm(tst))
        Results[r, 1:11] <- c(ind,tempdf$ECOREG_use[i],tempdf$ECOREG_use[j], #lk,
                              tempdf$NResp[i],
                              tempdf$NResp[j],tempdf$Estimate[i],
                              tempdf$Estimate[j], diff, stderr,tst,pval)
      }
    }
  }
}


###########
## ECOREGION ONLY
# 7/9/19 - SIZE ADJUSTED NLA 2012 dataset
write.csv(Results,"C:/Users/EFergus/OneDrive - Environmental Protection Agency (EPA)/a_Water_Level/Analysis/NLA_weighted_calculations/R_output/NLA12/Group_means_tests_NLA12_SIZEADJ_ECOREG_ONLY_SCALED_09JUL19.csv" )
